1389: From review: - fix Categoryized type - move CategoryOrderComparator to CategoryDefinition.orderComparator so it can be used by both the service and tool - fix cray cray indenting
