+++ gradebookng/tool/src/java/org/sakaiproject/gradebookng/business/GradebookNgBusinessService.java
@@ -329,7 +329,7 @@ public class GradebookNgBusinessService {
 		}
 
 		// Sort by categoryOrder
-		Collections.sort(rval, new CategoryOrderComparator());
+		Collections.sort(rval, CategoryDefinition.orderComparator);
 
 		return rval;
 	}
@@ -1521,7 +1521,7 @@ public class GradebookNgBusinessService {
 
 		final GradebookInformation settings = this.gradebookService.getGradebookInformation(gradebook.getUid());
 
-		Collections.sort(settings.getCategories(), new CategoryOrderComparator());
+		Collections.sort(settings.getCategories(), CategoryDefinition.orderComparator);
 
 		return settings;
 	}
@@ -1832,26 +1832,6 @@ public class GradebookNgBusinessService {
 					.append(subtotal1, subtotal2)
 					.toComparison();
 
-		}
-	}
-
-	/**
-	 * Comparator class for sorting categories by categoryOrder
-	 */
-	public class CategoryOrderComparator implements Comparator<CategoryDefinition> {
-		@Override
-		public int compare(final CategoryDefinition c1, final CategoryDefinition c2) {
-			if (c1.getCategoryOrder() == null && c2.getCategoryOrder() == null) {
-				return c1.getName().compareTo(c2.getName());
-			}
-			if(c1.getCategoryOrder() == null) {
-				return -1;
-			}
-			if(c2.getCategoryOrder() == null) {
-				return 1;
-			}
-
-			return c1.getCategoryOrder().compareTo(c2.getCategoryOrder());
 		}
 	}
 }
